Nesting with timeout.
--HG--
branch : feature

+++ src/libs/vlayout/vcontour.h
@@ -66,8 +66,8 @@ public:
     void             SetContour(const QVector<QPointF> &contour);
     QVector<QPointF> GetContour() const;
 
-    quint32 GetShift() const;
-    void    SetShift(quint32 shift);
+    qreal GetShift() const;
+    void  SetShift(qreal shift);
 
     int  GetHeight() const;
     void SetHeight(int height);
